merge fixed into main

See merge request zurihaqi/car-rent!2

+++ src/main/java/zul_ipin/car_rent/service/implementation/BrandServiceImplementation.java
@@ -1,13 +1,15 @@
 package zul_ipin.car_rent.service.implementation;
 
-import jakarta.persistence.criteria.CriteriaBuilder;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.jpa.domain.Specification;
 import org.springframework.stereotype.Service;
 import zul_ipin.car_rent.model.Brand;
 import zul_ipin.car_rent.repository.BrandRepository;
 import zul_ipin.car_rent.service.BrandService;
+import org.springframework.data.domain.Pageable;
+import zul_ipin.car_rent.utils.specification.BrandSpesification;
 
-import java.util.List;
 
 @Service
 @RequiredArgsConstructor
@@ -20,24 +22,36 @@ public class BrandServiceImplementation implements BrandService {
     }
 
     @Override
-    public List<Brand> getAll(){
-        return brandRepository.findAll();
+    public Page<Brand> getAll(Pageable pageable, String name){
+        Specification<Brand> spec = BrandSpesification.getSpesification(name);
+        return brandRepository.findAll(spec, pageable);
     }
 
     @Override
     public Brand getOne(Integer id){
-        return brandRepository.findById(id).orElse(null);
+        return brandRepository.findById(id).orElseThrow(() ->
+                new RuntimeException("Brand with id " + id + " not found!"));
     }
 
     @Override
-    public Brand update(Brand request){
-        Brand brand = this.getOne(request.getId());
-        brand.setName(request.getName());
-        return brandRepository.save(brand);
+    public Brand update(Integer id, Brand request){
+        if(brandRepository.findById(id).isEmpty()) {
+            throw new RuntimeException("Brand with id " + id + " not found!");
+        }
+        else {
+            Brand brand = this.getOne(id);
+            brand.setName(request.getName());
+            return brandRepository.save(brand);
+        }
     }
 
     @Override
     public void delete(Integer id){
-        brandRepository.deleteById(id);
+        if(brandRepository.findById(id).isEmpty()) {
+            throw new RuntimeException("Brand with id " + id + " not found!");
+        }
+        else{
+            brandRepository.deleteById(id);
+        }
     }
 }
